dashboard

+++ app/Http/Controllers/EkrafController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Model\Umkm;
 use App\Model\Product;
-use App\Model\Exhibition;
 use Yajra\Datatables\Datatables;
 use Response;
 use Excel;
@@ -34,7 +33,7 @@ class EkrafController extends Controller
              "business_form"=>$request->business,
              "address"=>$request->address,
              "year"=>$request->year,
-             "logo" =>  url("/")."/gambar/umkm/".$newName,
+             "logo" =>  $newName,
              "phone"=>$request->phone
          );
 
@@ -60,13 +59,17 @@ class EkrafController extends Controller
      $umkm = Umkm::find($id);
      if($umkm->delete()){
        if(isset($umkm->logo)){
-        $url_image = explode("/", $umkm->logo);
-        if(count($url_image)>0){
-        $path = public_path()."/gambar/umkm/".$url_image[count($url_image)-1];
-        if(file_exists($path)){
-          unlink($path);
-          }
+        if($umkm->logo!="recyle.png" ||$umkm->logo!="masyarakat.png"||$umkm->logo!="kontruksi.png"||$umkm->logo!="jacket.png"||$umkm->logo!="industri.png"||$umkm->logo!="food.png"||$umkm->logo!="farming.png"||$umkm->logo!="deliverybox.png"){
+           $oldPath = $umkm->logo;
+            if(isset($oldPath)){
+            $path = public_path()."/gambar/umkm/".$oldPath;
+            if(file_exists($path)){
+              unlink($path);
+              }
+            }
         }
+       
+
        }
      }
      
@@ -83,8 +86,8 @@ class EkrafController extends Controller
 
 
     public  function editUMKM(Request $request){
-        $umkm= Umkm::find($request->umkm_id);  
-    	$oldPath=explode("/", $umkm->logo);
+      $umkm= Umkm::find($request->umkm_id);  
+    	$oldPath=$umkm->logo;
       $oldLogo=$umkm->logo;
     	$logo=$request->logo;
     	$year=$request->year;
@@ -118,8 +121,8 @@ class EkrafController extends Controller
       $umkm->phone=$phone;
       $umkm->RT=$rt;
       $umkm->RW=$rw;
-      $umkm->female_labor=$female_labor;
-      $umkm->male_labor=$male_labor;
+      $umkm->female_labour=$female_labour;
+      $umkm->male_labour=$male_labour;
       $umkm->female_employe=$female_employe;
       $umkm->male_employe=$male_employe;
 
@@ -134,7 +137,7 @@ class EkrafController extends Controller
 		        $umkm->logo= url("/gambar/umkm/".$newName);
 		        $check = $umkm->save();
 		        if($check && !empty($oldLogo)){
-		          $path = public_path()."/gambar/umkm/".$oldPath[count($oldPath)-1];
+		          $path = public_path()."/gambar/umkm/".$oldPath;
 		          if(file_exists($path)){
 		             unlink($path);
 		          }
@@ -147,7 +150,7 @@ class EkrafController extends Controller
     }
 
     public function showUMKM(){
-    	  return  Datatables::of(Umkm::select("id","umkm_name","phone","business_form","logo","address","year","owner"))
+    	  return  Datatables::of(Umkm::select("id","umkm_name","phone","business_sector","logo","address","year","owner"))
        ->addColumn('action','
             <center>
               <a href="{{ URL::to("/umkm/$id") }}"   data-id="{{ $id }}" data-original-title="Detail" class="detail btn btn-primary detail-umkm" >
@@ -183,7 +186,7 @@ class EkrafController extends Controller
                 $produkData=array(
                      "name"=>$request->name,
                      "umkm_id"=>$request->umkm_id,
-                     "photo"=> url("/")."/gambar/umkm/produk/".$newName,
+                     "photo"=> $newName,
                      "price"=>$request->price
                   );
 
@@ -231,18 +234,17 @@ class EkrafController extends Controller
     $produk->name=$request->name;
     $produk->price=$request->price;
     $photo=$request->photo;
-    $oldPath=explode("/", $produk->logo);
-  
+    $oldPath=$produk->logo;
     
         if(isset($photo)){
             $ext = $photo->getClientOriginalExtension();
             $newName = rand(100000,1001238912).".".$ext;
             $gambar_path = public_path()."/gambar/umkm/produk";
             $photo->move($gambar_path, $newName);
-            $produk->photo= url("/gambar/umkm/produk/".$newName);
+            $produk->photo= $newName;
             $check = $produk->save();
             if($check && isset($oldPath)){
-              $path = public_path()."/gambar/umkm/produk/".$oldPath[count($oldPath)-1];
+              $path = public_path()."/gambar/umkm/produk/".$oldPath;
               if(file_exists($path)){
                  unlink($path);
               }
@@ -262,122 +264,12 @@ class EkrafController extends Controller
       } catch (Exception $e) {
               return Redirect::back();
       }
+          
+    
+     
     }
 
 
 
-    public function pameran (){
-        return view("exhibition");
-    }
-
-
-
-     public function showPameran(){
-        return  Datatables::of(Exhibition::select("id","exhibition_name","description","start_date","end_date","price","poster"))
-       ->addColumn('action','
-            <center>
-             <a href="#"   data-id="{{ $id }}" data-original-title="Edit" class="edit btn btn-success edit-pameran"  data-toggle="modal" data-target="#ajax-crud-modal">
-            <i class="fa fa-edit"></i>
-              </a>
-            
-           <a href="#" class="btn btn-danger" onclick="deletePameran({{$id}})"><i class="fa fa-trash"></i></a>
-              </center>')
-        ->rawColumns(['action'])
-        ->addIndexColumn()
-        ->make(true);
-    }
-
-
-    public function createPameran(Request $request){
-        $poster=$request->poster;
-
-        $pameranData;
-
-           if(isset($poster)){
-                $ext = $poster->getClientOriginalExtension();
-                $newName = rand(100000,1001238912).".".$ext;
-                $gambar_path = public_path()."/gambar/umkm/pameran";
-                $poster->move($gambar_path, $newName);
-                $pameranData=array(
-                     "exhibition_name"=>$request->pameran,
-                     "start_date"=>$request->mulai,
-                     "end_date"=>$request->akhir,
-                     "description"=>$request->deskripsi,
-                     "poster"=> $newName,
-                     "location"=>$request->location,
-                     "price"=>$request->price
-                  );
-
-           }else{
-               $pameranData=array(
-                     "exhibition_name"=>$request->pameran,
-                     "start_date"=>$request->mulai,
-                     "end_date"=>$request->akhir,
-                     "description"=>$request->deskripsi,
-                     "location"=>$request->location,
-                     "price"=>$request->price
-                  );
-           }
-
-              $exhibition=Exhibition::create($pameranData);
-              return $exhibition;
-
-    }
-
-
-    public function deletePameran($id){
-     $pameran = Exhibition::find($id);
-     if($pameran->delete()){
-       if(isset($pameran->poster)){
-        $poster = $pameran->poster;
-        if(isset($poster)){
-        $path = public_path()."/gambar/umkm/pameran".$poster;
-        if(file_exists($path)){
-          unlink($path);
-          }
-        }
-       }
-      return Response::json($pameran);
-
-     }
-   }
-
-
-   public function profilePameran($id){
-     $pameran = Exhibition::find($id);
-     return $pameran;
-   }
-
-
-   public function editPameran(Request $request){
-      $pameran=Exhibition::find($request->pameran_id);
-      $pameran->exhibition_name=$request->pameran;
-      $pameran->start_date=$request->mulai;
-      $pameran->end_date=$request->akhir;
-      $pameran->description=$request->deskripsi;
-      $pameran->location=$request->location;
-      $pameran->price=$request->price;
-      $poster=$request->poster;
-      $oldPoster=$pameran->poster; 
-     if(isset($poster)){
-            $ext = $logo->getClientOriginalExtension();
-            $newName = rand(100000,1001238912).".".$ext;
-            $gambar_path = public_path()."/gambar/umkm/pameran";
-            $pameran->move($gambar_path, $newName);
-            $pameran->poster=$newName;
-            $check = $pameran->save();
-            if($check && !empty($oldPoster)){
-              $path = public_path()."/gambar/umkm/pameran".$oldPoster;
-              if(file_exists($path)){
-                 unlink($path);
-              }
-            }
-          return Response::json($pameran);
-        }
-
-          $pameran->save();         
-        return Response::json($pameran);
-
-   }
 
 }
